Post event website (#1193)
* Setup structure for post event

* Start of designs

* Gif added

* Fix the thing

* Sizing
