+++ sass/event.scss
@@ -1191,6 +1191,55 @@ footer {
     }
 }
 
+// idk lol
+footer#end-footer {
+    position: absolute;
+}
+
+// Event things
+#end-main {
+    display: flex;
+    flex-direction: row;
+    // border: 3px solid yellow;
+    height: calc(100vh - #{$nav-height});
+    margin: 0 auto;
+    margin-top: $nav-height / 2;
+    width: 90%;
+    bottom: 0;
+
+    #thank-you-container, #ricky-container {
+        // border: 3px solid yellow;
+        display: flex;
+        flex-direction: column;
+        justify-content: flex-end;
+    }
+
+    #thank-you-container {
+        width: 40%;
+        // border: 3px solid yellow;
+
+        h1 {
+            font-size: 4em;
+        }
+
+        a {
+            color: $red;
+            margin-top: 20px;
+            font-size: 1.3em;
+
+            &:hover {
+                text-decoration: underline;
+            }
+        }
+    }
+
+    #ricky-container {
+        width: 60%;
+        // border: 3px solid yellow;
+        overflow: hidden; // Beeg gif
+    }
+}
+
 
 // Media queries for mobile
